added new class ModalDialogExamples.cs (extracted tests from ApiExamples.cs)
