+++ src/Coypu.AcceptanceTests/Coypu.AcceptanceTests.csproj
@@ -79,6 +79,7 @@
     <Compile Include="Examples\EnterTextExamples.cs" />
     <Compile Include="Examples\FindExamples.cs" />
     <Compile Include="Examples\HasExamples.cs" />
+    <Compile Include="Examples\ModalDialogExamples.cs" />
     <Compile Include="Examples\WindowExamples.cs" />
     <Compile Include="Examples\WithinExamples.cs" />
     <Compile Include="FindAllWithPredicateAndRetries.cs" />
